integration_tests/pyproject.toml
[project]
name = "integration-tests"
version = "0.1.0"
description = "Integration tests for decomp.me requiring cromper and Django backend"
requires-python = ">=3.10"
dependencies = [
    "pytest>=8.0.0",
    "pytest-django>=4.9.0",
    "pytest-env>=1.1.0",
    "requests>=2.32.3",
    "responses>=0.25.3",
    # All backend dependencies
    "django>=5.1.3",
    "django-cors-headers>=4.6.0",
    "django-environ>=0.11.2",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.10",
    "jwt>=1.3.1",
    "PyGithub>=2.5.0",
    "drf-extensions>=0.7.1",
    "tzdata==2024.2",
    "Pillow==11.0",
    "html-json-forms>=1.1.1",
    "django-resized>=1.0.3",
    "django-cleanup>=9.0.0",
    "sentry-sdk>=2.19.0",
    "django-session-timeout>=0.1.0",
    "setuptools>=75.6.0",
    "attrs>=24.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "django-stubs[compatible-mypy]>=5.2.5",
    "djangorestframework-stubs>=3.15.1",
    "types-requests>=2.32.0.20241016",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-v --tb=short --strict-markers"
testpaths = ["."]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "integration: marks tests as integration tests (requires cromper + Django)",
]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["__pycache__"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "PIE", # flake8-pie
    "PT",  # flake8-pytest-style
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["PT009", "PT027"]  # Allow pytest fixtures and assertions

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]
disable_error_code = ["var-annotated", "import-untyped"]

[tool.django-stubs]
django_settings_module = "decompme.settings"
